# groot/rtree: introduce FormulaFunc