This CL introduces a new way to create Formulae: FormulaFunc.
here is the comparison wrt Formula:

```
## ref: with Formula.
## new: with FormulaFunc.
$> benchstat ref.txt new.txt
name                   old time/op    new time/op    delta
FormulaFunc/f0/Eval-4     346ns ± 2%     238ns ± 2%  -31.30%  (p=0.000 n=18+19)
FormulaFunc/f0/Func-4     346ns ± 3%     230ns ± 2%  -33.50%  (p=0.000 n=19+20)
FormulaFunc/f1/Eval-4     359ns ± 2%     312ns ± 2%  -13.03%  (p=0.000 n=19+17)
FormulaFunc/f1/Func-4     365ns ± 2%     304ns ± 2%  -16.70%  (p=0.000 n=20+19)
FormulaFunc/f2/Eval-4     361ns ± 2%     314ns ± 2%  -13.06%  (p=0.000 n=20+18)
FormulaFunc/f2/Func-4     368ns ± 2%     304ns ± 1%  -17.29%  (p=0.000 n=20+18)
FormulaFunc/f3/Eval-4     384ns ± 2%     313ns ± 2%  -18.63%  (p=0.000 n=17+19)
FormulaFunc/f3/Func-4     386ns ± 1%     306ns ± 2%  -20.78%  (p=0.000 n=16+20)

name                   old alloc/op   new alloc/op   delta
FormulaFunc/f0/Eval-4     72.0B ± 0%     40.0B ± 0%  -44.44%  (p=0.000 n=20+20)
FormulaFunc/f0/Func-4     64.0B ± 0%     40.0B ± 0%  -37.50%  (p=0.000 n=20+20)
FormulaFunc/f1/Eval-4     72.0B ± 0%     48.0B ± 0%  -33.33%  (p=0.000 n=20+20)
FormulaFunc/f1/Func-4     64.0B ± 0%     48.0B ± 0%  -25.00%  (p=0.000 n=20+20)
FormulaFunc/f2/Eval-4     72.0B ± 0%     48.0B ± 0%  -33.33%  (p=0.000 n=20+20)
FormulaFunc/f2/Func-4     64.0B ± 0%     48.0B ± 0%  -25.00%  (p=0.000 n=20+20)
FormulaFunc/f3/Eval-4     72.0B ± 0%     48.0B ± 0%  -33.33%  (p=0.000 n=20+20)
FormulaFunc/f3/Func-4     64.0B ± 0%     48.0B ± 0%  -25.00%  (p=0.000 n=20+20)

name                   old allocs/op  new allocs/op  delta
FormulaFunc/f0/Eval-4      3.00 ± 0%      2.00 ± 0%  -33.33%  (p=0.000 n=20+20)
FormulaFunc/f0/Func-4      2.00 ± 0%      2.00 ± 0%     ~     (all equal)
FormulaFunc/f1/Eval-4      3.00 ± 0%      2.00 ± 0%  -33.33%  (p=0.000 n=20+20)
FormulaFunc/f1/Func-4      2.00 ± 0%      2.00 ± 0%     ~     (all equal)
FormulaFunc/f2/Eval-4      3.00 ± 0%      2.00 ± 0%  -33.33%  (p=0.000 n=20+20)
FormulaFunc/f2/Func-4      2.00 ± 0%      2.00 ± 0%     ~     (all equal)
FormulaFunc/f3/Eval-4      3.00 ± 0%      2.00 ± 0%  -33.33%  (p=0.000 n=20+20)
FormulaFunc/f3/Func-4      2.00 ± 0%      2.00 ± 0%     ~     (all equal)
```

all stats:
```
name                   time/op
Formula/f0/Eval-4      346ns ± 2%
Formula/f0/Func-4      346ns ± 3%
Formula/f1/Eval-4      359ns ± 2%
Formula/f1/Func-4      365ns ± 2%
Formula/f2/Eval-4      361ns ± 2%
Formula/f2/Func-4      368ns ± 2%
Formula/f3/Eval-4      384ns ± 2%
Formula/f3/Func-4      386ns ± 1%
FormulaFunc/f0/Eval-4  243ns ± 2%
FormulaFunc/f0/Func-4  235ns ± 2%
FormulaFunc/f1/Eval-4  315ns ± 0%
FormulaFunc/f1/Func-4  307ns ± 2%
FormulaFunc/f2/Eval-4  316ns ± 2%
FormulaFunc/f2/Func-4  309ns ± 1%
FormulaFunc/f3/Eval-4  317ns ± 2%
FormulaFunc/f3/Func-4  310ns ± 2%

name                   alloc/op
Formula/f0/Eval-4      72.0B ± 0%
Formula/f0/Func-4      64.0B ± 0%
Formula/f1/Eval-4      72.0B ± 0%
Formula/f1/Func-4      64.0B ± 0%
Formula/f2/Eval-4      72.0B ± 0%
Formula/f2/Func-4      64.0B ± 0%
Formula/f3/Eval-4      72.0B ± 0%
Formula/f3/Func-4      64.0B ± 0%
FormulaFunc/f0/Eval-4  40.0B ± 0%
FormulaFunc/f0/Func-4  40.0B ± 0%
FormulaFunc/f1/Eval-4  48.0B ± 0%
FormulaFunc/f1/Func-4  48.0B ± 0%
FormulaFunc/f2/Eval-4  48.0B ± 0%
FormulaFunc/f2/Func-4  48.0B ± 0%
FormulaFunc/f3/Eval-4  48.0B ± 0%
FormulaFunc/f3/Func-4  48.0B ± 0%

name                   allocs/op
Formula/f0/Eval-4       3.00 ± 0%
Formula/f0/Func-4       2.00 ± 0%
Formula/f1/Eval-4       3.00 ± 0%
Formula/f1/Func-4       2.00 ± 0%
Formula/f2/Eval-4       3.00 ± 0%
Formula/f2/Func-4       2.00 ± 0%
Formula/f3/Eval-4       3.00 ± 0%
Formula/f3/Func-4       2.00 ± 0%
FormulaFunc/f0/Eval-4   2.00 ± 0%
FormulaFunc/f0/Func-4   2.00 ± 0%
FormulaFunc/f1/Eval-4   2.00 ± 0%
FormulaFunc/f1/Func-4   2.00 ± 0%
FormulaFunc/f2/Eval-4   2.00 ± 0%
FormulaFunc/f2/Func-4   2.00 ± 0%
FormulaFunc/f3/Eval-4   2.00 ± 0%
FormulaFunc/f3/Func-4   2.00 ± 0%
```