@sbinet 2020-05-04
from https://github.com/go-hep/hep/issues/647#issuecomment-623410102:
> Nice (that was fast!!)! So it seems we can get another 15-30% improvement, meaning roughly twice slower than the direct access via `rtree.ReadVar`. 
>
> There is something I don't understand though: now, there is no more "interpreted" operation, right? >So why don't we get the same time than the `rtreeReadVar`? If you just load branches, and call a go function on it, it should be the same as ... loading branches and calling a go function on the values :) (I guess I am missing something but ...)
> 
> Anyway, it is definitely cool !

I haven't made a detailed analysis but my hypothesis is that we are now hitting the reflect toll gate:
- to locate+load values from the `rtree.ReadVar`
- dereference the `rtree.ReadVar.Value` pointers, copy them and pass the values to the user-provided function
- call the user provided function, get the output
- pass it back to the user

all of this is done via `reflect` to be able to handle all types of input parameters and all types of output values.

when/if "generics" land into Go (perhaps in 1-2 years, see [there](https://go.googlesource.com/proposal/+/master/design/go2draft-contracts.md)), a fair amount of `reflect` code could probably go away from `FormulaFunc`.