@rmadar 2020-05-04
OK, I have tested it and it works as expected: we are now at a factor ~2.0-2.3 from the direct reading:
 + `ReadVar`: 3.0 ms/kEvt
 + `Formula`: 8.5 ms/kEvt 
 + `FormulaFunc`: 6.7 ms/kEvt

That's quite a big step wrt the very first shot!

I may have a small suggestion, if that's easy: it would be to print the name of the variable that could not be found (it is useful for dumb user like me!). For now, we have:
```
could not create formulaFunc: 
 rtree: could not create FormulaFunc: 
  rtree: could not find all needed ReadVars
```