@sbinet 2020-05-05
should be fixed w/ https://github.com/go-hep/hep/pull/693