@rmadar 2020-05-12
I have a question on the way `FormulaFunc` works. Let's assume I want to defined branch-based variable and code-based variable in the same way (for *e.g.*, `"top_pt*3"` or `0.87`). I would do (thanks to a kind of wrapper) something like:

```go
// Case 1
f, err := r.FormulaFunc(
    []string{"top_pt"}),
    func (pt float64) float64 {return pt*3},
)

// Case 2
f, err := r.FormulaFunc(
    []string{},
    func () float64 {return 0.87},
)
```

My question is: do we have some of the expensive `reflect` calls when the slice of strings is empty like in case 2?