@sbinet 2020-05-12
I expect the difference between the 2 versions to be minimal.
I don't expect the biggest performance drop to come from the number of arguments (and passing them to the user function) but rather from the call to the user function, through `reflect`.

of course, I may be mistaken :)