@rmadar 2020-05-12
OK thanks. After thinking a bit, I just changed my wrapper to return the simple function `f` if not info from the tree is needed, and a `r.FormulaFunc(names, f).Func().(T)` otherwise. It's more clean anyway.